comportamento do Ctrl+C
- Mudança na função close para pegar a thread que está ocupando a região crítica

+++ include/enviar.h
@@ -3,6 +3,6 @@
 
 #include "conexao.h"
 
-int enviar(int sockfd, void *dados, size_t tamanho);
+ssize_t enviar(int sockfd, void *dados, size_t tamanho);
 
 #endif //ENVIAR_H
